+++ apps/web/modules/paypal/runtime/components/ApplePayButton.vue
@@ -4,6 +4,7 @@
 
 <script lang="ts" setup>
 import type { PayPalAddToCartCallback } from '../types';
+import { usePayPal } from '../composables/usePayPal';
 
 const { initialize, config, processPayment, getTransactionInfo } = useApplePay();
 const { getCurrentScript } = usePayPal();
