+++ apps/web/modules/paypal/runtime/components/GooglePayButton.vue
@@ -13,6 +13,7 @@
 <script lang="ts" setup>
 import type { PayPalAddToCartCallback } from '../types';
 import { SfLoaderCircular } from '@storefront-ui/vue';
+import { usePayPal } from '../composables/usePayPal';
 
 const {
   initialize,
