+++ apps/web/modules/paypal/runtime/composables/usePayPal/index.ts
@@ -1,3 +1,2 @@
 export * from './usePayPal';
-export * from './types';
 export * from './usePayPalVisibility';
