+++ apps/web/modules/paypal/runtime/types/index.ts
@@ -1,1 +1,2 @@
 export * from './PayPalComponentTypes';
+export * from './PayPalComposableTypes';
